(parse) Split parser tests into multiple suites
Since Jest parallelizes test suites (i.e. files), we can speed up parser
tests by testing different expression types at the same time.

They're not quite as fast as the old lexer-only tests were using Mocha +
Chai, but there's many more tests now so perhaps that's to be expected.
Plus they still finish in about one second :shrug:
